src/screens/Register.js: moved to register dir, updated imports
